Restructure packaging in rule.evaluation component
