+++ components/rule-mgt/org.wso2.carbon.identity.rule.evaluation/pom.xml
@@ -84,15 +84,11 @@
                         </Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Private-Package>
-                            org.wso2.carbon.identity.rule.evaluation.internal,
-                            org.wso2.carbon.identity.rule.evaluation.core,
-                            org.wso2.carbon.identity.rule.evaluation.service.impl,
+                            org.wso2.carbon.identity.rule.evaluation.internal
                         </Private-Package>
                         <Export-Package>
                             !org.wso2.carbon.identity.rule.evaluation.internal,
-                            !org.wso2.carbon.identity.rule.evaluation.core,
-                            !org.wso2.carbon.identity.rule.evaluation.service.impl,
-                            org.wso2.carbon.identity.rule.evaluation.*;
+                            org.wso2.carbon.identity.rule.evaluation.api.*;
                             version="${carbon.identity.package.export.version}",
                         </Export-Package>
                         <Import-Package>
