+++ components/user-mgt/org.wso2.carbon.identity.user.pre.update.password.action/pom.xml
@@ -124,9 +124,9 @@
                             org.wso2.carbon.identity.certificate.management.service; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.certificate.management.exception; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.certificate.management.model; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.rule.evaluation.model; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.rule.evaluation.exception; version="${carbon.identity.package.import.version.range}",
-                            org.wso2.carbon.identity.rule.evaluation.provider; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.identity.rule.evaluation.api.model; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.identity.rule.evaluation.api.exception; version="${carbon.identity.package.import.version.range}",
+                            org.wso2.carbon.identity.rule.evaluation.api.provider; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.user.action.api.model; version="${carbon.identity.package.import.version.range}",
                             org.wso2.carbon.identity.user.action.api.service; version="${carbon.identity.package.import.version.range}"
                         </Import-Package>
